requirements.txt
allure-pytest==2.8.16
allure-python-commons==2.8.16
apipkg==1.5
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
DateTime==4.3
execnet==1.7.1
idna==2.10
iniconfig==1.0.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
PyMySQL==0.10.0
pyparsing==2.4.7
pytest==5.4.3
pytest-forked==1.3.0
pytest-html==2.1.1
pytest-metadata==1.10.0
pytest-rerunfailures==9.0
pytest-xdist==1.34.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-mock==1.8.0
six==1.15.0
toml==0.10.1
urllib3==1.25.9
wcwidth==0.2.5
zope.interface==5.1.0

yagmail~=0.11.224
jsonpath~=0.82
loguru~=0.5.1